Move the control message reading to another class
